requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Django==4.2.2
django-bootstrap4==23.1
django-widgets-improved==1.5.0
et-xmlfile==1.1.0
openpyxl==3.1.2
pdfminer.six==20221105
pycparser==2.21
python-magic==0.4.27
python-magic-bin==0.4.14
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
asgiref==3.7.2
beautifulsoup4==4.12.2
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Django==4.2.2
django-bootstrap4==23.1
django-upload-validator==1.1.6
django-widgets-improved==1.5.0
et-xmlfile==1.1.0
gunicorn==20.1.0
openpyxl==3.1.2
pdfminer.six==20221105
psycopg2-binary==2.9.6
pycparser==2.21
python-magic==0.4.27
python-magic-bin==0.4.14
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
